Added new function argument for building consensus skeleton with jackknife
Jackknife is as method that randomly selects samples from a fraction of
the original data (undersampling). For that, we need an additional
function argument so that the user can specify this fraction. In order
to leave bootstrap (maybe it will be used in the future), bootstrap
will be replaced as resampling method if this new parameter is set,
which means jackknife will be used instead. I also fixed a typo I
did in previous commits (bootstraping -> bootstrapping).