Added safety checks for another consensus skeleton parameter
For the jackknife method, the fraction (proportion) of the original
data that will be sampled, a process of undersampling, can not be
over 100% (the whole data) or 0% (no data). The value 100 is
accepted for it is the default and indicates that the method used
should be bootstrapping. If set to any other accepted value,
that is, 0 > proportionToUndersample < 100, jackknife will be
used instead.