Implemented jackknife method for consensus skeleton estimation
Jackknife method will be used if the proportionToUndersample
parameter is set to any value other than 100 (without running
into the restrictions imposed by safety checks). This method will
randomly selects N samples from the original data, N being
the total number of samples minus the proportion of this number
according to proportionToUndersample.

If we have 100 samples and proportionToUndersample is set to 80,
we will randomly select 80 samples from the original data.