Changed loop for consensus skeleton feature
Initially, I wanted to make the loop more efficient by using the
last skeleton inferred both for the final network and the
consensus network. For bootstrapping that was easy, since I could
simply use the original data, but for jackknife it was tricky.
In a previous commit I wrote a workaround to make it work. However,
the code wasn't very clear and by having the last skeleton used for
both the consensus and the final network, added bias to the
similarity analysis between the consensus skeleton and the final
network. Therefore I fixed both issues (workaround+bias) by adding
a new iteration (nSkeletons+1, from now on) to the loop.