Updated report file on consensus skeleton
I updated the report file with results for both nSkeletons=10 and
100. Besides, I stored statically the outputs so that the render
of the R Markdown file is quicker and allows me to modify it more
easily. The miic calls were done with a seed in order to be able
to reproduce the same results dynamically at a late time, if
required.