Evaluation of parameters used for consensus skeleton inference
I added two new sections to the report.

(1) I wrote a script to infer the consensus skeleton from
hematoData for several different values of the three parameters
(number of skeletons, frequency of occurrence to be added to
the consensus skeleton and fraction of original data that will be
randomly sampled for the jackknife algorithm) that are used for
the consensus skeleton inference. I used the same seed to
allow for comparison among the results.

(2) I also added a section to briefly describe the data used in
this analyses (hematoData).