Fixed bug for skeletons from resamplings
When I rewrote the calls to infer skeletons in the previous commit,
I accidently passed the original dataset for all calls, instead of
the resampled subsets for the skeletons to build the consensus
skeleton. Besides, there is a bug that is crashing the confidence
cut when I choose only the skeletons from resampled data to get
their edges filtered. It's for now fixed because of line 542. This
should be a temporary fix.