Started working on benchmarking with insurance dataset
For now, I'm doing with the insurance dataset what I did before
with the hematoData dataset. The goal is to see if the results
previously seen are comparable. With that in hand, I will improve
the benchmarking code to compare with the known real network of
the insurance dataset and increase the number of parameters used
in the benchmarking.