Updated benchmark function and report (WIP)
I started working on the benchmark function with a workflow more
similar to what I recently discussed with Prof. Isambert. We're
going to obtain the F-score for the skeleton inferred from the
data with edge filtering and compare with the F-score for the
skeleton inferred from the data with the consensus feature. The
pseudocode is almost finished in this commit, though more must be
done.