Added last updated timestamp to report
I'm uploading this report to my website whenever I update
something so that Prof. Isambert can always have access to the
most up to date version of the report. Thus, a timestamp showing
when the document was last updated at the head of the document
is needed so that he can easily see that there are changes he
still has not seen.