Updated function call for simulations
I updated the scripts and the function to run simulations comparing
the consensus skeleton approach with the edge filtering approach
with the ALARM, INSURANCE and BARLEY dataset. I also organized the
folders in my git repository so that output files, which are big
because of images, are symbolically in the repo folder but actually
backed up in my Dropbox.