Updated report on benchmarking details
I updated the benchmarking section of the report to make it much
clearer on how I performed the simulations and the benchmarking,
including source code for the specific functions, how I used the
scripts and so on. I also added information on the real networks
that I used for the benchmarking.