Changed default knitr (report) options
Warning and extra output messages should be omitted to as minimum
verbose as possible so that only what I really want to put in the
report is shown. Due to some very large scripts, I also decided to
hide by default codes. In order to see them, one gotta click on
the button "code".