Fixed color of points in the 3D interactive plot
The color of the points should refer to the difference between the
Fscore of the two methods but it was kind of random. By reading
the documentation of the package, I didn't find how to make it
behave the way I expected it to behave so I changed to a fixed
color (red). The Fscore (among other parameters) can be viewed
by hovering the mouse on the point.