Fixed unrecognized global functions causing a note
Since ggplot2 is only suggested, its namespace is not fully imported
(and it should not, we don't want to force users to install ggplot to
use our package). Every function which namespace was not imported has
to be called using the prefix pkg::f() or it causes a note during the
CRAN checks. It (marginally) beats having to create a separate package.