Allow non-exact p-value in Spearman correlation test.
During the computation of Spearman rank correlation, the p-value cannot
be computed exactly when ties are present in the data, but will instead
be approximated, which gives a warning if exact is set to TRUE (by
default). This warning is of little importance for the miic result, but
the presence of a lot of them can hide other more useful warnings.
Ref: ?stats::cor.test