Fix bug in the computation of reduced levels.
This is a minor bug of an edge case:
When zi_list.size() > 1, and a candidate Z_i leads to reduced number of
levels of X or Y being equal to or less than 1 (because of NA), and the
next candidate Z_(i+1) contains no NA, then the reduced number of levels
of X and Y w.r.t. Z_(i+1) will be erroneously set to keep those of Z_i,
leading to incorrect calculation of score.