Use threadprivate and raw pointer for li_alloc_ptr
The change from thread_local to threadprivate is done to fix check
error on 32-bit Windows, which is likely to be related to the bug of
mingw when handling thread_local variables.
Ref: https://sourceforge.net/p/mingw-w64/bugs/527/,
https://github.com/dmlc/xgboost/pull/2994

Then, the change from std::unique_ptr to raw pointer is to detour the
bug of GCC not working well with threadprivate for non POD type
variable, giving variable "declared 'threadprivate' after first use"
compile-time error.
Ref: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=27557